fix circular deps bug
